test/dspy/tests/requirements.txt
pytest>=7.0.0
pytest-mock>=3.10.0
dspy-ai
weaviate-client
python-dotenv
litellm
replicate
transformers
torch 